refactor(tui): split sprint-detail-view into focused sibling files
- Approach B: orchestrator stays at the canonical path so importers and
  the 496-LOC test stay unchanged; siblings under sprint-detail-internals/
- Extracted: header-card / task-summary (tasks list) / ticket-list /
  action-bar / shortcuts + focus-list / shared-prose / attempt-card /
  field-editors / use-sprint-bundle
- Orchestrator down from 1320 -> 309 LOC; every file under 350
